Last.fm recent track list no longer uses Tracks, but Scrobbles instead
This allows us, for example, to have duplicates in the list of recent
tracks.

This temporarily stops app from stopping/starting autoloading of tracks
that are or are not now playing, but it should not have been doing that
to begin with. Need to investigate further action to make this work
properly.